requirements.txt
crewai
python-dotenv
pyyaml
google-generativeai
langchain-google-genai
langchain-openai  